lib/simbricks/axil_manager: fix `step_apply()` sometimes being skipped
When `AXILManager::pending_` only contained a single operation and that one was removed from `AXILManager::pending_` in `step()` on completion, there were still some signals that needed to be reset to their idle values. The temp variables were correctly modified to do so in `step()`, however, `step_apply()` was skipped since `AXILManager::pending_` was now empty.